Added more propmts for the single test task
Gave the user the option to build a coverage report when running a
single test in an attempt to make the tests run quicker.  There is only
about a 2 second gain between running all tests and a single test while
generating a report.  Without a report, running a single test is about
30 seconds quicker than running all tests with a report.